% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_playlists.R
\name{ovp_example_playlist}
\alias{ovp_example_playlist}
\title{Example playlists}
\usage{
ovp_example_playlist(choice)
}
\arguments{
\item{choice}{numeric: which data file to return?
\itemize{
\item{1 - Pipe attacks by Poland in the Poland - Iran men's World League match, June 2017}
}}
}
\value{
A data.frame containing the playlist.
}
\description{
Example playlists
}
\examples{
\dontrun{
 ## read example playlist
 ply <- ovp_example_playlist(1)

 ## start shiny app
 ovp_shiny(playlist = ply)
}

}
\seealso{
\code{\link{ovp_shiny}}
}
