% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{ovp_shiny}
\alias{ovp_shiny}
\title{Launch the Shiny app}
\usage{
ovp_shiny(
  playlist,
  video_server = NULL,
  launch_browser = TRUE,
  mobile = FALSE,
  host = "127.0.0.1",
  ...
)
}
\arguments{
\item{playlist}{data.frame: a playlist as returned by \code{\link[ovideo]{ov_video_playlist}}}

\item{video_server}{string: the method used to serve local video files, either "lighttpd" or "servr". If it is \code{NULL}, no local server will be started and only remote video files can be used}

\item{launch_browser}{logical: if \code{TRUE}, launch the app in the system's default web browser (passed to \code{\link[shiny]{runApp}}'s \code{launch.browser} parameter)}

\item{mobile}{logical: if \code{TRUE}, launch the app's mobile UI version}

\item{host}{string: shiny host. Defaults to "127.0.0.1"}

\item{...}{: additional parameters passed to the UI and server functions}
}
\description{
Launch the Shiny app
}
\examples{
\dontrun{
 ## read example playlist
 ply <- ovp_example_playlist(1)

 ## start shiny app
 ovp_shiny(playlist = ply)
}
}
